aarch64 headers & new interrupts vectors

+++ inc/arch/aarch64/aarchutils.h
@@ -2,6 +2,12 @@
 #define _AARCH64_H_
 
 #include "def/typedef.h"
+#include "arm64.h"
+#include "atomic.h"
+#include "barrier.h"
+#include "irqflags.h"
+#include "regs.h"
+#include "types.h"
 
 #define __AARCH64__ 1
 
@@ -10,6 +16,6 @@ void put32(int, int);
 
 void memzero(void *, int);
 void delay(void);
-int64 get_el(void);
+uint get_el(void);
 
 #endif
